Separated the initialization logic from the program's logic
